@@ -1,7 +1,6 @@
 package main
 
 import (
-	"io/ioutil"
 	"os"
 	"syscall"
 	"testing"
@@ -51,7 +50,7 @@ func TestReadConfig(t *testing.T) {
 			DNSConfig{},
 		},
 	} {
-		tmpfile, err := ioutil.TempFile("", "acmedns")
+		tmpfile, err := os.CreateTemp("", "acmedns")
 		if err != nil {
 			t.Error("Could not create temporary file")
 		}
@@ -99,7 +98,7 @@ func TestGetIPListFromHeader(t *testing.T) {
 }
 
 func TestFileCheckPermissionDenied(t *testing.T) {
-	tmpfile, err := ioutil.TempFile("", "acmedns")
+	tmpfile, err := os.CreateTemp("", "acmedns")
 	if err != nil {
 		t.Error("Could not create temporary file")
 	}
@@ -118,7 +117,7 @@ func TestFileCheckNotExists(t *testing.T) {
 }
 
 func TestFileCheckOK(t *testing.T) {
-	tmpfile, err := ioutil.TempFile("", "acmedns")
+	tmpfile, err := os.CreateTemp("", "acmedns")
 	if err != nil {
 		t.Error("Could not create temporary file")
 	}
